+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:enableOnBackInvokedCallback="true">
 
         <activity
-            android:name="uk.org.pentlandscouts.scoreboard_app.ui.navapp.ScoreboardActivity"
+            android:name="uk.org.pentlandscouts.scoreboard_app.ui.ScoreboardActivity"
             android:exported="true"
             android:label="ScoreBoard App"
             android:theme="@style/Theme.ScoreboardApp">
