--- app/src/main/java/uk/org/pentlandscouts/scoreboard_app/ui/ScaffoldExampleActivity.kt
@@ -1,207 +1,0 @@
-package uk.org.pentlandscouts.scoreboard_app.ui
-
-
-//noinspection UsingMaterialAndMaterial3Libraries
-import android.annotation.SuppressLint
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-//noinspection UsingMaterialAndMaterial3Libraries
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material3.BottomAppBar
-import androidx.compose.material3.DrawerState
-import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FabPosition
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults.topAppBarColors
-import androidx.compose.material3.rememberDrawerState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.example.scoreboard_app.R
-import com.example.scoreboard_app.R.*
-
-
-class ScaffoldSampleActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            ScaffoldRootView {
-                onBackPressed()
-            }
-        }
-    }
-
-
-}
-
-@Composable
-private fun ScaffoldRootView(backPressed: () -> Unit) {
-     ScaffoldSample()
-      ScaffoldWithTopBar(backPressed)
-    ScaffoldWithBottomMenu()
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@Composable
-fun ScaffoldWithTopBar(backPressed: () -> Unit) {
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = {
-                    Row {
-                        Text(
-                            text = "Title 1",
-                            fontSize = 30.sp,
-                            color = Color.Black
-                        )
-                        Text(
-                            text = " Title 2",
-                            fontSize = 30.sp,
-                            color = Color.Black
-                        )
-
-                    }
-                },
-                navigationIcon = {
-                    IconButton(onClick = {}) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, "backIcon")
-                    }
-                }
-            )
-        }, content = {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(Color(0xff8d6e63)),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Content of the page",
-                    fontSize = 30.sp,
-                    color = Color.Black
-                )
-            }
-
-        }
-    )
-}
-
-@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@Composable
-fun ScaffoldWithBottomMenu() {
-    Scaffold(bottomBar = { BottomBar() }
-    ) {
-        //content area
-        Box(
-            modifier = Modifier
-                .background(Color.White)
-                .fillMaxSize()
-
-        )
-
-        NavTopBar.TopBarNav()
-    }
-}
-
-@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ScaffoldSample() {
-    val scaffoldState = rememberScaffoldState(rememberDrawerState(DrawerValue.Closed))
-    Scaffold(
-
-        topBar = {
-            TopAppBar(
-                title = { NavTopBar.TopBarNav() },
-                colors = topAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.primaryContainer,
-                    titleContentColor =  MaterialTheme.colorScheme.primary ,
-                ),
-            )
-        },
-        floatingActionButtonPosition = FabPosition.End,
-        floatingActionButton = {
-            FloatingActionButton(onClick = {}) {
-                Icon(imageVector = Icons.Default.Add, contentDescription = "fab icon")
-            }
-        },
-
-        content = {
-            //if (selectedContent == "ScoreBoard")
-                ScoreBoardView(LocalContext.current)
-           // if (selectedContent == "AddScore")
-             //   AddScoreView(context = LocalContext.current)
-
-
-
-        },
-        bottomBar = { BottomAppBar { Text("Bottom App Bar") } }
-    )
-}
-
-fun rememberScaffoldState(rememberDrawerState: DrawerState) {
-
-}
-
-@Composable
-fun BottomBar() {
-    val selectedIndex = remember { mutableIntStateOf(0) }
-    BottomNavigation(elevation = 10.dp) {
-        BottomNavigationItem(icon = {
-            Icon(imageVector = Icons.Default.Favorite,"")
-        },
-            label = { Text(text = "Home") }, selected = (selectedIndex.value == 0), onClick = {
-                selectedIndex.intValue = 0
-            })
-        BottomNavigationItem(icon = {
-            Icon(imageVector = Icons.Default.Favorite,"")
-        },
-            label = { Text(text = "Favorite") }, selected = (selectedIndex.value == 1), onClick = {
-                selectedIndex.intValue = 1
-            })
-        BottomNavigationItem(icon = {
-            Icon(imageVector = Icons.Default.Person, "")
-        },
-            label = { Text(text = "Profile") }, selected = (selectedIndex.value == 2), onClick = {
-                selectedIndex.intValue = 2
-            })
-    }
-}
-
-@Preview(showSystemUi = true, showBackground = true)
-@Composable
-fun ScaffoldPreview() {
-    ScaffoldRootView {
-
-    }
-}
