+++ app/src/main/java/uk/org/pentlandscouts/scoreboard_app/ui/ScoreboardActivity.kt
@@ -1,14 +1,10 @@
-package uk.org.pentlandscouts.scoreboard_app.ui.navapp
+package uk.org.pentlandscouts.scoreboard_app.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.TopAppBar
@@ -16,18 +12,13 @@ import androidx.compose.material3.TopAppBarDefaults.topAppBarColors
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import kotlinx.coroutines.launch
 import uk.org.pentlandscouts.scoreboard_app.theme.Purple40
 import uk.org.pentlandscouts.scoreboard_app.theme.ScoreboardAppTheme
-import uk.org.pentlandscouts.scoreboard_app.ui.data
-import uk.org.pentlandscouts.scoreboard_app.ui.getScoreBoardData
-import uk.org.pentlandscouts.scoreboard_app.ui.navapp.components.ScoreboardTabRow
-import uk.org.pentlandscouts.scoreboard_app.ui.navapp.topbar.TopBarScreen
+import uk.org.pentlandscouts.scoreboard_app.ui.components.ScoreboardTabRow
+import uk.org.pentlandscouts.scoreboard_app.ui.topbar.TopBarScreen
 
 class ScoreboardActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
