+++ app/src/main/java/uk/org/pentlandscouts/scoreboard_app/ui/activities/Activities.kt
@@ -1,26 +1,21 @@
-package uk.org.pentlandscouts.scoreboard_app.ui.navapp.teams
+package uk.org.pentlandscouts.scoreboard_app.ui.activities
 
 import android.content.Context
 import android.widget.Toast
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Accessibility
 import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Person
 import androidx.compose.material3.Card
 import androidx.compose.material3.Icon
 import androidx.compose.material3.OutlinedButton
@@ -31,63 +26,40 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import org.json.JSONArray
-import uk.org.pentlandscouts.scoreboard_app.data.TeamItem
-import uk.org.pentlandscouts.scoreboard_app.ui.navapp.ScoreboardViewModel
-import uk.org.pentlandscouts.scoreboard_app.util.ResourceUtils
+import uk.org.pentlandscouts.scoreboard_app.data.ActivityItem
+import uk.org.pentlandscouts.scoreboard_app.ui.ScoreboardViewModel
 
 @Composable
-fun TeamsScreen(viewModel: ScoreboardViewModel, context: Context)
+fun ActivitiesScreen(viewModel: ScoreboardViewModel, context: Context)
 {
-    var teamsData:JSONArray = JSONArray()
-    var teamsList: List<TeamItem> = ArrayList()
+
+    val activitiesData: JSONArray
+    var activitiesList: List<ActivityItem>
     val isLoading by viewModel.isLoading.collectAsStateWithLifecycle()
 
     when(isLoading) {
         ScoreboardViewModel.State.Loading -> CircularProgressIndicator() // Show loading
 
         else -> {
-            teamsList = ArrayList()
-            teamsData = viewModel.teamsData
-            for (i in 0 until teamsData.length()) {
-                val row = teamsData.getJSONObject(i)
-                teamsList = teamsList +
-                        TeamItem(
+            activitiesList = ArrayList()
+            activitiesData = viewModel.activities
+            for (i in 0 until activitiesData.length()) {
+                val row = activitiesData.getJSONObject(i)
+                activitiesList = activitiesList +
+                        ActivityItem(
                             row.get("name").toString(),
-                            row.get("id").toString().toInt(),
-                            row.get("logo").toString().toString(),
+                            row.get("description").toString(),
+                            row.get("maxScore").toString().toInt(),
+                            row.get("zone").toString(),
+                            row.get("id").toString().toInt()
                         )
             }
-
-            Spacer(modifier = Modifier.height(9.dp))
-
-            Row(
-                Modifier
-                    .padding(5.dp)
-                    .background(Color.White)
-                    .fillMaxWidth(),
-                verticalAlignment = Alignment.Top,
-                horizontalArrangement = Arrangement.Start,
-            )
-            {
-                Icon(Icons.Default.Person, contentDescription = "Teams")
-                Text(text = "Teams", fontWeight = FontWeight.Bold, color = Color.Black)
-            }
-
-            Row()
-            {
-                Column()
-                {
-                    Text(text = "Teams")
-                }
-            }
-
 
             LazyVerticalGrid(
                 columns = GridCells.Fixed(1),
@@ -95,47 +67,48 @@ fun TeamsScreen(viewModel: ScoreboardViewModel, context: Context)
                     .padding(5.dp)
                     .background(Color.White)
             ) {
-                // Display Scoreboard Items
-                items(teamsList.size) {
 
+
+                // Display Scoreboard Items
+                items(activitiesList.size) {
 
                     Card(
                         onClick = {
                             // inside on click we are displaying the toast message.
                             Toast.makeText(
                                 context,
-                                "Show " + teamsList[it].teamName + " Team",
+                                "Show " + activitiesList[it].name + " `Details`..",
                                 Toast.LENGTH_SHORT
                             ).show()
                         },
                         // on below line we are adding padding from our all sides.
                         modifier = Modifier.padding(8.dp),
                     ) {
-                        // on below line we are creating a column on below sides.
+
                         Row(
                             modifier = Modifier.background(Color.White)
                         )
                         {
-                            Spacer(modifier = Modifier.width(35.dp))
+                            Spacer(modifier = Modifier.width(10.dp))
                             Text(
-                                text = "Logo",
+                                text = "Zone",
                                 fontWeight = FontWeight.Bold,
                                 color = Color.Black
                             )
-                            Spacer(modifier = Modifier.width(50.dp))
+                            Spacer(modifier = Modifier.width(25.dp))
                             Text(
-                                text = "Team Name",
+                                text = "Activity Name",
                                 fontWeight = FontWeight.Bold,
                                 color = Color.Black
                             )
                             Spacer(modifier = Modifier.width(20.dp))
                             Text(
-                                text = "ID",
+                                text = "Max Score",
                                 fontWeight = FontWeight.Bold,
                                 color = Color.Black
                             )
 
-                            Spacer(modifier = Modifier.width(40.dp))
+                            Spacer(modifier = Modifier.width(20.dp))
                             Text(
                                 text = "Edit",
                                 fontWeight = FontWeight.Bold,
@@ -143,6 +116,7 @@ fun TeamsScreen(viewModel: ScoreboardViewModel, context: Context)
                             )
                             Spacer(modifier = Modifier.width(50.dp))
                         }
+                        // on below line we are creating a column on below sides.
                         Row(
                             Modifier
                                 .fillMaxSize()
@@ -153,53 +127,39 @@ fun TeamsScreen(viewModel: ScoreboardViewModel, context: Context)
                         )
                         {
                             // on below line we are creating image for our grid view item.
-
-                            Image(
-                                // on below line we are specifying the drawable image for our image.
-
-                                //Get the resourceId from MipMap
-                                //var resID = ResourceUtils.getResourceIdFromMipMap(scoreBoardList[it].teamName)
-                                painter = painterResource(
-                                    id = ResourceUtils.getResourceIdFromMipMap(
-                                        teamsList[it].logo
-                                    )
-                                ),
-                                contentDescription = "Team Icon",
-                                modifier = Modifier
-                                    .height(60.dp)
-                                    .width(60.dp)
-                                    .padding(5.dp)
+                            Text(
+                                text = activitiesList[it].zone,
+                                color = Color.DarkGray
                             )
+
                             Spacer(modifier = Modifier.height(9.dp))
 
                             Text(
-                                text = teamsList[it].teamName,
-                                //scoreBoardList[it].teamName,
+                                text = activitiesList[it].name,
                                 modifier = Modifier.padding(4.dp),
                                 color = Color.DarkGray
                             )
 
-                            Spacer(modifier = Modifier.height(9.dp))
+                            Spacer(modifier = Modifier.width(20.dp))
 
                             Text(
-                                text = teamsList[it].id.toString(),
+                                text = activitiesList[it].maxScore.toString(),
                                 color = Color.DarkGray
                             )
 
-
-
+                            Spacer(modifier = Modifier.width(40.dp))
                             OutlinedButton(
                                 onClick = {
                                     Toast.makeText(
                                         context,
-                                        "Navigate to Edit  Team: " + teamsList[it].teamName,
+                                        "Navigate to Edit Activity: " + activitiesList[it].name,
                                         Toast.LENGTH_SHORT
                                     ).show()
                                 },
                                 contentPadding = PaddingValues(5.dp),
                             )
                             {
-                                Icon(Icons.Default.Edit, contentDescription = "Edit")
+                                Icon(Icons.Default.Edit, contentDescription = "Edit Activity")
                             }
                         }
                     }
@@ -207,12 +167,12 @@ fun TeamsScreen(viewModel: ScoreboardViewModel, context: Context)
             }
         }
     }
+
 }
 
-@Composable
 @Preview
-fun TeamsPreview()
+@Composable
+fun ActivitiesScreenPrev()
 {
-    TeamsScreen(viewModel = viewModel(), context = LocalContext.current)
+    ActivitiesScreen(viewModel = viewModel(), context = LocalContext.current)
 }
-
